Reorganize code to a better structure
