+++ src/tokenizer.cpp
@@ -1,15 +1,18 @@
 #include <string>
 #include <iostream>
-#include <list>
-#include <unordered_set>
-#include <unordered_map>
-#include <regex>
 
 
 #include "global_variables.hpp"
 #include "read_bib.hpp"
 #include "tokenizer.hpp"
+#include "string_utils.hpp"
 
+
+Token::Token (Tag input_tag, std::string input_value)
+{
+    tag = input_tag;
+    value = input_value;
+}
 
 Token buid_token(Tag type_of_token,
                 std::string::iterator begin_of_lexeme,
@@ -20,37 +23,22 @@ Token buid_token(Tag type_of_token,
     return token;
 }
 
-
-
-int count_char(std::string &string, char chr)
+std::list<Token> tokenizer()
 {
-    std::string::iterator current_char = string.begin();
-    int n_of_occurrences = 0;
-    while (current_char != string.end())
-    {
-        if (*current_char == chr)
-        {
-            n_of_occurrences++;
-        }
+    std::list<Token> tokens;
+    std::vector<SubString> bib_entries = collect_bib_entries(bibparser::bib_file);
+    std::for_each(bib_entries.begin(), bib_entries.end(), &trim_entry);
 
-        current_char++;
+    for (SubString entry : bib_entries)
+    {
+        parse_entry(entry);
     }
 
-    return n_of_occurrences;
+    return tokens;
 }
 
-void find_first_position(std::string &string, char chr)
-{
-    std::string::iterator it = string.begin();
-    while (it != string.end())
-    {
-        if (*it == chr) {
-            break;
-        }
 
-        it++;
-    }
-}
+
 
 
 std::vector<SubString> collect_bib_entries(std::string &file)
@@ -125,6 +113,23 @@ void trim_entry(SubString &entry)
 }
 
 
+
+void parse_entry(SubString entry)
+{
+    std::string::iterator end = entry.end;
+
+    SubString entry_type = find_entry_type(entry);
+    SubString entry_attrs = {entry_type.end, end};
+    SubString entry_identifier = find_entry_identifier(entry_attrs);
+    print_substring(entry_attrs);
+
+
+    //std::cout << std::string(entry_identifier.begin, entry_identifier.end) << std::endl;
+
+}
+
+
+
 SubString find_entry_identifier(SubString attrs)
 {
     std::string::iterator current_char = attrs.begin;
@@ -161,73 +166,5 @@ SubString find_entry_type(SubString entry)
     }
 
     return {begin, current_char + 1};
-}
-
-
-void print_substring(SubString substring)
-{
-    std::string s = std::string(substring.begin, substring.end);
-    std::cout << s << std::endl;
-}
-
-
-void parse_entry(SubString entry)
-{
-    std::string::iterator end = entry.end;
-
-    SubString entry_type = find_entry_type(entry);
-    SubString entry_attrs = {entry_type.end, end};
-    SubString entry_identifier = find_entry_identifier(entry_attrs);
-    print_substring(entry_attrs);
-
-
-    //std::cout << std::string(entry_identifier.begin, entry_identifier.end) << std::endl;
-
-}
-
-
-
-std::list<Token> tokenizer()
-{
-    std::list<Token> tokens;
-    std::vector<SubString> bib_entries = collect_bib_entries(bibparser::bib_file);
-    std::for_each(bib_entries.begin(), bib_entries.end(), &trim_entry);
-
-    for (SubString entry : bib_entries)
-    {
-        parse_entry(entry);
-    }
-
-    return tokens;
-}
-
-
-
-
-
-bool is_white_space(char chr)
-{
-    bool white_space = chr == ' ' || chr == '\t' || chr == '\r' || chr == '\n';
-    return white_space ? true : false;
-}
-
-bool is_digit(char chr)
-{
-    return find_in_set(chr, NUMBERS);
-}
-
-bool is_letter(char chr)
-{
-    return find_in_set(chr, LETTERS);
-}
-
-bool is_line(char chr)
-{
-    return find_in_set(chr, LINES);
-}
-
-bool find_in_set(char chr, const std::unordered_set<char>& set)
-{
-    return set.find(chr) != set.end() ? true : false;
 }
